requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-filter==23.3
django-cors-headers==4.3.1
Pillow==10.1.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
python-decouple==3.8
whitenoise==6.6.0
drf-spectacular==0.26.5
django-extensions==3.2.3
gunicorn==21.2.0
# numpy==1.24.3
# opencv-python-headless==4.8.1.78
# face-recognition==1.3.0
# faiss-cpu==1.7.4
django-debug-toolbar
